+++ src/components/modals/create-coopertaion-modal/index.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { useRouter } from 'next/navigation';
+import { usePathname, useRouter } from 'next/navigation';
 
 import { Button } from '@/components/ui/button';
 import {
@@ -12,15 +12,34 @@ import {
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { useModal } from '@/hooks/useModal';
+import { useAuthStore } from '@/store/authStore';
+import { PATHS, STORAGE_KEY_AUTH } from '@/utils';
 
 export function CreateCoopertaionModal() {
+  const { setAuth, getAuth } = useAuthStore();
   const router = useRouter();
+  const pathname = usePathname();
   const [docName, setDocName] = useState('');
   const { isOpen, onClose, type } = useModal();
   const isModalOpen = isOpen && type === 'createCooperation';
 
   const handleCreate = async () => {
-    // 发送 POST 请求
+    let auth = getAuth();
+
+    if (!auth.access_token) {
+      const storagedAuth = localStorage.getItem(STORAGE_KEY_AUTH);
+
+      if (storagedAuth) {
+        auth = JSON.parse(storagedAuth);
+
+        setAuth(auth);
+      }
+    }
+
+    if (!auth.access_token) {
+      router.push(`${PATHS.LOGIN}?redirect=${pathname}`);
+    }
+
     const response = await fetch(`${process.env.NEXT_PUBLIC_SERVER_URL}/api/v1/document/create`, {
       method: 'POST',
       headers: {
